% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcs_url.R
\name{gcs_url}
\alias{gcs_url}
\title{Generate image URL}
\usage{
gcs_url(bucket_name, glider_deployment, image_type, directory_name, image_name)
}
\arguments{
\item{bucket_name}{string; name of GCP bucket}

\item{glider_deployment}{string; name of glider deployment as glider-YYYYmmdd}

\item{image_type}{string; image type, aka name of intermediary path folder}

\item{directory_name}{string; directory name}

\item{image_name}{string; name of the shadowgraph image}
}
\value{
The authenticated URL, as a string
}
\description{
Generate the GCS Authenticated URL for the user-specified shadowgraph image
}
\details{
Glues the inputs together to generate the proper GCS Authenticated URL.
}
\examples{
gcs_url("amlr-gliders-imagery-raw-dev", "amlr08-20220513", "images", 
        "Dir0000", "SG01 20220513-190538-011.jpg")

}
