% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glider.deployments}
\alias{glider.deployments}
\title{ESD glider deployment data}
\format{
\subsection{\code{glider.deployments}}{

A data frame:
\describe{
\item{deployment}{Glider deployment code}
\item{project}{ESD glider project code; either SANDIEGO or FREEBYRD}
\item{year}{Year of deployment, matches GCS path}
}
}
}
\usage{
glider.deployments
}
\description{
ESD glider deployments and project information. Incomplete list. Used for
generating lists for the shade Shiny app and Google Cloud Storage (GCS) URLs
}
\keyword{datasets}
