% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_string.R
\name{gcs_url}
\alias{gcs_url}
\alias{region_image_name_parse}
\title{Generate image URL}
\usage{
gcs_url(bucket_name, glider_deployment, image_type, directory_name, image_name)

region_image_name_parse(image_name, bucket_name, image_type = "")
}
\arguments{
\item{bucket_name}{string; name of GCP bucket}

\item{glider_deployment}{string; name of glider deployment as glider-YYYYmmdd}

\item{image_type}{string; image type, aka name of intermediary path folder.
Must be one of: 'images-ffPCG', 'images-imgff', 'jpgorig-regions'}

\item{directory_name}{string; directory name}

\item{image_name}{string; name of the shadowgraph image. Note that for
\code{region_image_name_parse()}, this must be the name of a region image (see
Examples section)}

\item{z}{string; name of GCP bucket}
}
\value{
string
}
\description{
These functions perform different parts of image name processing:
\itemize{
\item \code{gcs_url()} Glues the inputs together to generate the proper GCS
Authenticated URL, returning this URL as a string
\item \code{region_image_name_parse()}: Parses a region file name, and outputs a string
with the name of the user-specificed version of the parent image
Generate the GCS Authenticated URL for the user-specified shadowgraph image
}
}
\examples{
gcs_url("amlr-gliders-imagery-raw-dev", "amlr08-20220513", "images", 
        "Dir0000", "SG01 20220513-190538-011.jpg")
gcs_url("amlr-gliders-imagery-proc-dev", "amlr08-20220513", "images-ffPCG", 
        "Dir0000", "SG01 20220513-190538-011-ffPCG.png")
        
region_image_name_parse("SG01 20220510-213508-001.jpg-out0.jpg", 
                         "amlr-gliders-imagery-proc-dev", "images-ffPCG")
region_image_name_parse("SG01 20220510-213508-001.jpg-out0.jpg", 
                         "amlr-gliders-imagery-raw-dev")

}
