% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matrix.R
\name{make_dispersions}
\alias{make_dispersions}
\title{Make dispersions dataframe}
\usage{
make_dispersions(Q, focal_cell, potentials, verbose = FALSE)
}
\arguments{
\item{Q}{Logistic regression quadrature scheme, as generated by make_quadrature}

\item{focal_cell}{Focal cell type (character) whose interactions we're modeling}

\item{potentials}{Named list of potential functions, each taking a distance and returning a scalar}

\item{verbose}{Logical flag for verbosity}
}
\value{
Sparse matrix of dispersion features
}
\description{
Make dispersions dataframe
}
