% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matrix.R
\name{make_rbfs}
\alias{make_rbfs}
\title{Make radial basis functions}
\usage{
make_rbfs(max_dist, n_basis_functions = 6, basis_function_sigma = 8)
}
\arguments{
\item{max_dist}{maximum distance to consider}

\item{n_basis_functions}{number of basis functions}

\item{basis_function_sigma}{standard deviation for all basis functions}
}
\value{
list of named radial basis functions
}
\description{
Make radial basis functions
}
