% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_SHADE_gq}
\alias{run_SHADE_gq}
\title{Generate predictions from fitted SHADE model.}
\usage{
run_SHADE_gq(data_gq, draws_fit, ...)
}
\arguments{
\item{data_stan}{list of arguments needed for SHADE model
objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>}
}
\value{
Model fit
}
\description{
Generate predictions
}
\seealso{
Other models: 
\code{\link{run_SHADE_model}()}
}
\concept{models}
