% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_SHADE_model}
\alias{run_SHADE_model}
\title{Fit the SHADE model.}
\usage{
run_SHADE_model(data_stan, method = "sample", ...)
}
\arguments{
\item{data_stan}{list of arguments needed for SHADE model}

\item{method}{Character string specifying the inference method: "sample" for MCMC sampling (default) or "variational" for variational inference}

\item{...}{Named arguments to the `sample()` or `variational()` method of CmdStan model
objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>}
}
\value{
Model fit
}
\description{
Fit the SHADE Stan model and return fit
}
\seealso{
Other models: 
\code{\link{run_SHADE_gq}()}
}
\concept{models}
