% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHADE-package.R
\docType{package}
\name{SHADE-package}
\alias{SHADE}
\alias{SHADE-package}
\title{SHADE: Spatial Hierarchical Asymmetry via Directional Estimation}
\description{
A framework for modeling directional spatial interactions between cell types in multiplexed imaging data
using a hierarchical Bayesian model. SHADE estimates spatial interaction curves (SICs) at image, patient,
and cohort levels, capturing asymmetric and multiscale spatial organization.
}
\keyword{internal}
