% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{shade}
\alias{shade}
\title{shade}
\usage{
shade(...)
}
\arguments{
\item{...}{passed to \code{\link[shiny]{shinyApp}}}
}
\description{
Runs shade, the shiny-shadow-fixer R Shiny App
}
\details{
Shiny app that allows users to generate the parent GCS URL for a shadowgraph
region image, and fix a VIAME CSV file using user-provided thresholds
}
\examples{
if (interactive()) shade()

}
