% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{smooth_density_fft}
\alias{smooth_density_fft}
\title{Perform 2D FFT-Based Density Smoothing (Simple Version)}
\usage{
smooth_density_fft(X, kernel_func, resolution = 128)
}
\arguments{
\item{X}{A point pattern (`ppp` object) or density grid (`im` object)}

\item{kernel_func}{A function that computes kernel values dynamically}

\item{resolution}{Resolution of the density grid if `X` is a `ppp` object}
}
\value{
A smoothed density estimate as an `im` object
}
\description{
Perform 2D FFT-Based Density Smoothing (Simple Version)
}
