% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viame.csv.metadata}
\alias{viame.csv.metadata}
\title{VIAME CSV metadata}
\format{
An object of class \code{character} of length 2.
}
\source{
\url{https://kitware.github.io/dive/DataFormats/},
\url{https://viame.readthedocs.io/en/latest/section_links/detection_file_conversions.html}
}
\usage{
viame.csv.metadata
}
\description{
Components of line 2 of a VIAME CSV file, i.e. the static metadata
information. This data is used by \code{viame_csv_metadata()} to construct the
full metadata info
}
\keyword{datasets}
