% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viame.csv.names}
\alias{viame.csv.names}
\title{VIAME CSV column names}
\format{
An object of class \code{character} of length 11.
}
\source{
\url{https://kitware.github.io/dive/DataFormats/},
\url{https://viame.readthedocs.io/en/latest/section_links/detection_file_conversions.html}
}
\usage{
viame.csv.names
}
\description{
The column names of a VIAME CSV file downloaded from VIAME-Web
}
\keyword{datasets}
