% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viame_csv_fixer.R
\name{read_viame_csv}
\alias{read_viame_csv}
\alias{viame_csv_metadata}
\alias{write_viame_csv}
\alias{fix_viame_csv}
\title{VIAME CSV fixer}
\usage{
read_viame_csv(file)

viame_csv_metadata()

write_viame_csv(x, file)

fix_viame_csv(x, threshold_base, threshold_df = NULL)
}
\arguments{
\item{file}{file path; see \code{\link[readr:read_delim]{readr::read_csv()}}}

\item{x}{tibble; with format the same as the output of \code{read_viame_csv()}}

\item{threshold_base}{numeric; base confidence threshold from VIAME-Web}

\item{threshold_df}{tibble; data frame of individual type thresholds from
VIAME-Web that are different from \code{threshold_base}. Must have two columns:
'class' and 'threshold'}
}
\value{
\itemize{
\item \code{read_viame_csv()}: tibble, output of \code{\link[readr:read_delim]{readr::read_csv()}}
\item \code{viame_csv_metadata()}: tibble with metadata to write to VIAME CSV
\item \code{write_viame_csv()}: see \code{\link[readr:write_delim]{readr::write_csv()}}
\item \code{fix_viame_csv()}: tibble of a fixed VIAME CSV
}
}
\description{
VIAME CSV fixer functions: read/write VIAME CSV files, and filter by thresholds
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
threshold.df <- data.frame(class = "surf", threshold = 0.89)
read.adv <- read_viame_csv("C:/Users/PinnipedData/Downloads/test-images-adv.csv")
fix.adv <- fix_viame_csv(read.adv, 0.02)
fix_viame_csv(read.adv, 0.02, threshold.df)

write_viame_csv(fix.adv, "C:/Users/PinnipedData/Downloads/tmp.csv")
read_viame_csv("C:/Users/PinnipedData/Downloads/tmp.csv")
\dontshow{\}) # examplesIf}
}
