% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matrix.R
\name{make_data}
\alias{make_data}
\title{Make regression dataframe from quadrature}
\usage{
make_data(Q, potentials, focal_cell = NULL, verbose = FALSE)
}
\arguments{
\item{Q}{Logistic regression quadrature scheme, as generated by make_quadrature}

\item{potentials}{Named list of potential functions, each taking a distance as input}

\item{focal_cell}{Focal cell type (character) for which the regression is computed}

\item{verbose}{Logical flag for verbosity (default: FALSE)}
}
\value{
A list containing the design matrix (`data`) with an intercept and dispersion features,
        and the response vector (`response`) indicating data (1) vs. dummy (0) points.
}
\description{
Make regression dataframe from quadrature
}
