% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_SHADE_model}
\alias{run_SHADE_model}
\title{Fit the SHADE model.}
\usage{
run_SHADE_model(data_stan, ...)
}
\arguments{
\item{data_stan}{list of arguments needed for SHADE model}

\item{...}{Named arguments to the `sample()` method of CmdStan model
objects: <https://mc-stan.org/cmdstanr/reference/model-method-sample.html>}
}
\value{
Model fit
}
\description{
Fit the SHADE Stan model and return fit
}
\concept{models}
