% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade-package.R
\docType{package}
\name{shade-package}
\alias{shade-package}
\title{Shiny Shadow Fixer}
\description{
shade: Shiny Shadow Fixer
}
\details{
R package/Shiny app for fixing VIAME CSV output files.
See \href{https://github.com/Kitware/dive/issues/1415}{this bug} for more details.
}
\author{
Sam Woodman \email{sam.woodman@noaa.gov}
}
\keyword{package}
