% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{shade}
\alias{shade}
\title{shade}
\usage{
shade(...)
}
\arguments{
\item{...}{passed to \code{\link[shiny]{shinyApp}}}
}
\value{
todo
}
\description{
Runs shade, the shiny-shadow-fixer R Shiny App
}
\details{
todo
}
\examples{
if (interactive()) shade()

}
